Restore proposed methods for PR
Due to name/directory change (mv saltstack -> salt) and subsequently
edits, git could not keep track that `Pkg.java` was the same file.
With this commit we restore the original modifications to the
`Pkg.java` file in the new location.
